feat: cambios

+++ routes/web.php
@@ -2,16 +2,6 @@
 
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
 
 Route::get('/', function () {
     return view('welcome');
@@ -21,17 +11,25 @@ Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
 
-// Specialty routes
-Route::get('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'index']);
+// Admin routes
+Route::middleware(['auth', 'admin'])->group(function () {
+    // Specialty routes
+    Route::get('/especialidades', [App\Http\Controllers\admin\SpecialtyController::class, 'index']);
+    Route::get('/especialidades/create', [App\Http\Controllers\admin\SpecialtyController::class, 'create']);
+    Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\admin\SpecialtyController::class, 'edit']);
+    Route::post('/especialidades', [App\Http\Controllers\admin\SpecialtyController::class, 'store']);
+    Route::put('/especialidades/{specialty}', [App\Http\Controllers\admin\SpecialtyController::class, 'update']);
+    Route::delete('/especialidades/{specialty}', [App\Http\Controllers\admin\SpecialtyController::class, 'destroy']);
 
-Route::get('/especialidades/create', [App\Http\Controllers\SpecialtyController::class, 'create']);
-Route::get('/especialidades/{specialty}/edit', [App\Http\Controllers\SpecialtyController::class, 'edit']);
-Route::post('/especialidades', [App\Http\Controllers\SpecialtyController::class, 'store']);
-Route::put('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'update']);
-Route::delete('/especialidades/{specialty}', [App\Http\Controllers\SpecialtyController::class, 'destroy']);
+    // Doctor routes
+    Route::resource('medicos', App\Http\Controllers\admin\DoctorController::class);
 
-// Doctor routes
-Route::resource('medicos', App\Http\Controllers\DoctorController::class);
+    // Patient routes
+    Route::resource('pacientes', App\Http\Controllers\admin\PatientController::class);
 
-// Patient routes
-Route::resource('pacientes', App\Http\Controllers\PatientController::class);
+});
+
+Route::middleware(['auth', 'doctor'])->group(function () {
+    Route::get('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'edit']);
+    Route::post('/horario', [App\Http\Controllers\Doctor\HorarioController::class, 'store']);
+});
